career_agent/requirements.txt
Flask
python-docx
pdfplumber
